requireTournamentOrganizer
